(feat): confirmation de suppression

+++ src/pages/AdminPage.jsx
@@ -59,6 +59,10 @@ function AdminPage() {
     };
 
     const handleDelete = async (productID) => {
+        if (!confirm('Êtes-vous sûr de vouloir supprimer ce produit ?')) {
+            return;
+        }
+
         try {
             const response = await fetch(`/api/deleteProduct`, {
                 method: 'DELETE',
